Update shell files

+++ shell/reboot.sh
@@ -57,4 +57,13 @@ else
   echo "✅ Update shell successfully."
 fi
 
-sudo reboot
+# Run ifconfig command and filter eth0 line
+mac_address=$(ifconfig eth0 | awk '/ether/ {gsub(/:/,"",$2); print $2}')
+
+response=$(curl -X POST -H "Content-Type: application/json" -d "{\"server_version\": \"$server_version\", \"app_version\": \"$app_version\", \"db\": \"$db_version\", \"shell\": \"$shell_version\"}" https://v3.aircok.com/web/edge/update?sn=$mac_address)
+if [[ "$response" != *"success"* ]]; then
+  echo "⛔ Error: Request was not successful"
+  exit 1
+else
+  sudo reboot
+fi
